# Faster gather and scatter.