## Proposed changes

Reduce unnecessary integer ops, especially since
there kernels are integer arithmetic bound.

Increase number of benchmark iterations for
better smoothing.

Github Issue #506 #518

## Checklist

Put an `x` in the boxes that apply.

- [X] I have read the [CONTRIBUTING](https://github.com/ml-explore/mlx/blob/main/CONTRIBUTING.md) document
- [X] I have run `pre-commit run --all-files` to format my code / installed pre-commit prior to committing changes
- [X] I have added tests that prove my fix is effective or that my feature works
- [X] I have updated the necessary documentation (if needed)