@vj-krish 2024-02-13
### Perf on M1 Max MBP (32-core GPU, 64GB RAM)

#### gather_bench (Before)

====================
X (100, 64), Indices (1000000,)
MLX: 3.128ms
PyTorch: 13.422ms
====================
X (100, 1024), Indices (100000,)
MLX: 4.917ms
PyTorch: 21.406ms
====================
X (4, 1000000), Indices ()
MLX: 0.270ms
PyTorch: 0.173ms

#### gather_bench (After)

(mlx2) vjkrish@Vijays-MacBook-Pro-3 mlx % python benchmarks/python/gather_bench.py
====================
X (100, 64), Indices (1000000,)
MLX: 2.232ms
PyTorch: 13.336ms
====================
X (100, 1024), Indices (100000,)
MLX: 3.434ms
PyTorch: 21.071ms
====================
X (4, 1000000), Indices ()
MLX: 0.277ms
PyTorch: 0.194ms