@vj-krish 2024-02-13
### Scatter Perf on M1 Max MBP (32-core GPU, 64GB RAM)

#### Before

X (1000000, 64), Indices (1000000,)
MLX: 6.627ms
PyTorch: 13.380ms

X (1000000, 64), Indices (1000000,)
MLX: 6.778ms
PyTorch: 14.002ms

X (100000, 64), Indices (100000,)
MLX: 2.381ms
PyTorch: 1.571ms

#### After

X (1000000, 64), Indices (1000000,)
MLX: 5.925ms
PyTorch: 13.205ms

X (1000000, 64), Indices (1000000,)
MLX: 6.045ms
PyTorch: 14.023ms

X (100000, 64), Indices (100000,)
MLX: 2.280ms
PyTorch: 1.522ms